Fix development audit issues
