+++ package.json
@@ -20,21 +20,18 @@
     "del": "^3.0.0",
     "execa": "^0.8.0",
     "gulp": "^4.0.1",
-    "gulp-jsdoc3": "^1.0.1",
+    "gulp-jsdoc3": "^2.0.0",
     "gulp-jshint": "^2.0.4",
-    "gulp-mocha": "^4.3.1",
+    "gulp-mocha": "^7.0.1",
     "gulp-sourcemaps": "^2.6.1",
-    "gulp-tslint": "^8.1.1",
-    "gulp-typescript": "^3.2.2",
-    "gulp-util": "^3.0.8",
     "jsdoc": "^3.3.2",
     "jshint": "^2.9.5",
     "make-dir": "^1.1.0",
     "merge2": "^1.1.0",
-    "mocha": "^3.5.3",
-    "mocha-jenkins-reporter": "^0.3.9",
+    "mocha": "^6.2.0",
+    "mocha-jenkins-reporter": "^0.4.2",
     "ncp": "^2.0.0",
-    "nyc": "^11.7.2",
+    "nyc": "^14.1.1",
     "pify": "^3.0.0",
     "run-sequence": "^2.2.0",
     "semver": "^5.5.0",
@@ -50,17 +47,6 @@
       "name": "Google Inc."
     }
   ],
-  "nyc": {
-    "include": [
-      "packages/grpc-health-check/health.js",
-      "packages/grpc-js/build/src/*",
-      "packages/grpc-native-core/index.js",
-      "packages/grpc-native-core/src/*.js",
-      "packages/proto-loader/build/src/*"
-    ],
-    "cache": true,
-    "all": true
-  },
   "scripts": {
     "test": "nyc gulp test && GRPC_DNS_RESOLVER=ares nyc gulp nativeTestOnly",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
